+++ src/Platform/Providers/RouteServiceProvider.php
@@ -45,20 +45,6 @@ class RouteServiceProvider extends ServiceProvider
                 ? $role->where('id', $value)->firstOrFail()
                 : $role->where('slug', $value)->firstOrFail();
         });
-
-        Route::bind('widget', function ($value) {
-            try {
-                $widget = app()->make(Crypt::decryptString($value));
-
-                abort_if(! is_a($widget, WidgetContractInterface::class), 403);
-
-                return $widget;
-            } catch (\Exception $exception) {
-                Log::alert($exception->getMessage());
-
-                abort(404, $exception->getMessage());
-            }
-        });
     }
 
     /**